requirements.txt
gymnasium
stable-baselines3
pyopengl
pyopengl_accelerate
numpy
